+++ Backend/CampusTradeSystem/CampusTrade.API/appsettings.json
@@ -8,7 +8,7 @@
         "Microsoft.EntityFrameworkCore": "Information",
         "Microsoft.EntityFrameworkCore.Database.Command": "Warning",
         "CampusTrade.API.Services.Cache": "Debug",
-        "CampusTrade.API.Services.BackgroundServices": "Information",
+        "CampusTrade.API.Services.Background": "Information",
         "CampusTrade.API.Controllers.CacheManagementController": "Information"
       }
     },
@@ -88,6 +88,7 @@
     "SenderEmail": "1870707155@qq.com",
     "SenderName": "校园交易系统",
     "EnableSsl": true
+  },
   "CacheSettings": {
     "MemoryCache": {
       "SizeLimit": 1024,
