--- Backend/CampusTradeSystem/EmailConsoleTest/appsettings.json
@@ -1,18 +1,0 @@
-{
-  "Email": {
-    "SmtpServer": "smtp.qq.com",
-    "SmtpPort": 587,
-    "Username": "1870707155@qq.com",
-    "Password": "gsqlerqxaryaefcf",
-    "SenderEmail": "1870707155@qq.com",
-    "SenderName": "校园交易系统",
-    "EnableSsl": true
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  }
-}
